+++ R/cc.R
@@ -7,19 +7,16 @@
 # -------------------------------------------------------------------------
 #' @param ...
 #'
-#' Unquoted names (separated by commas) that you wish to quote.
+#' Either unquoted names (separated by commas) that you wish to quote or a
+#' length one character vector you wish to split by whitespace.
 #'
 #' Empty arguments (e.g. third item in `one,two,,four`) will be returned as `""`.
 #'
+#' Character vectors not of length one are returned as is.
+#'
 #' @param .clip `[bool]`
 #'
-#' Should the code to generate the constructed character vector be copied to
-#' your system clipboard.
-#'
-#' Defaults to FALSE unless the option "imp.clipboard" is set to TRUE.
-#'
-#' Note that copying to clipboard requires the availability of package
-#' [clipr](https://cran.r-project.org/package=clipr).
+#' Not currently used.
 #'
 # -------------------------------------------------------------------------
 #' @return
@@ -34,31 +31,24 @@
 # -------------------------------------------------------------------------
 #' @importFrom utils capture.output
 #' @export
-cc <- function(..., .clip = getOption("imp.clipboard", FALSE)) {
-    res <- substitute(list(...))
-    # we use as.character rather than deparse as we simply want quoted names
-    res <- as.character(res[-1])
-    if (interactive() && isTRUE(.clip)) {
-        if (!requireNamespace("clipr", quietly = TRUE)) {
-            message("Unable to copy to clipboard: install 'clipr' for this functionality.")
-        } else if (suppressWarnings(clipr::clipr_available())) {
-            clipr::write_clip(capture.output(dput(res, control = "all")))
-        } else if (Sys.info()["sysname"] != "Linux" || !.last_ditch(res)) {
-            message("Unable to copy to clipboard.")
+cc <- function(..., .clip = FALSE) {
+    if (!isFALSE(.clip)) {
+        warning(
+            "`.clip` is not currently used by the ympes package. ",
+            "Please read the NEWS for version 1.5.0 for further details."
+        )
+    }
+
+    if(...length() == 1L && is.character(..1)) {
+        res <- ..1
+        if (length(res) == 1L) {
+            res <- strsplit(trimws(res), split = "[[:space:]]+", perl = TRUE)
+            res <- res[[1L]]
         }
+    } else {
+        res <- substitute(list(...))
+        # we use as.character rather than deparse as we simply want quoted names
+        res <- as.character(res[-1])
     }
     res
-}
-
-.last_ditch <- function(res) {
-    display <- Sys.getenv("DISPLAY")
-    if (display == "")
-        return(FALSE)
-    cmd <- sprintf("xclip -selection clipboard -i  -display %s", display)
-    con <- pipe(cmd, open = "w")
-    out <- try(writeChar(capture.output(dput(res, control = "all")), con))
-    close(con)
-    if (inherits(out, "try-error"))
-        return(FALSE)
-    TRUE
 }