+++ inst/tinytest/test_cc.R
@@ -2,3 +2,9 @@ expect_identical(
     cc(  dale, audrey, laura  , hawk, ),
     c("dale", "audrey", "laura", "hawk", "")
 )
+
+expect_identical(
+    cc("  dale audrey laura   hawk "),
+    c("dale", "audrey", "laura", "hawk")
+)
+